pyproject.toml
[project]
name = "federatedchannelcharting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "h5py>=3.10",
    "hydra-core>=1.3",
    "joblib>=1.3",
    "lightning>=2.2.0",
    "loguru>=0.7",
    "matplotlib>=3.8",
    "mlflow>=2.10",
    "numpy>=1.26",
    "omegaconf>=2.3",
    "opencv-python>=4.9",
    "optuna>=3.5",
    "optuna-dashboard>=0.15",
    "pandas>=2.2",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "rich>=13.7",
    "scikit-learn>=1.4",
    "scipy>=1.12",
    "seaborn>=0.13",
    "seed-everything>=0.1.0",
    "tensorboard>=2.15",
    "torch>=2.2.0",
    "torchaudio>=2.2.0",
    "torchinfo>=1.8",
    "torchmetrics>=1.3",
    "torchvision>=0.17.0",
    "tqdm>=4.66",
    "wandb>=0.16",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
